requirements.txt
PyYAML
Jinja2
pybtex